pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-seo_analyzer"
version = "0.1.0"
description = "Seo Analyzer environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.115.0",
    "openai>=1.40.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m seo_analyzer.server.app
server = "seo_analyzer.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["seo_analyzer", "seo_analyzer.server"]
package-dir = { "seo_analyzer" = ".", "seo_analyzer.server" = "server" }